Move the changes in the right files
