+++ offgridplanner/projects/forms.py
@@ -39,3 +39,16 @@ class OptionForm(ModelForm):
         model = Options
         fields = list(OPTIONS_LABELS)
         labels = OPTIONS_LABELS
+
+
+# Bachirou added
+class BoundForm(ModelForm):
+    class Meta:
+        model = Options
+        fields = ["latitude_min", "latitude_max", "longitude_min", "longitude_max"]
+        labels = {
+            "latitude_min": _("Latitude min"),
+            "latitude_max": _("Latitude max"),
+            "longitude_min": _("Longitude min"),
+            "longitude_max": _("Longitude max"),
+        }
