+++ offgridplanner/projects/models.py
@@ -18,6 +18,20 @@ class Options(models.Model):
     do_grid_optimization = models.BooleanField(default=True)
     do_es_design_optimization = models.BooleanField(default=True)
 
+    # import pdb; pdb.set_trace() # Add column in the DB and check the validation of the coord boundaries
+    latitude_min = models.FloatField(
+        default=4.2, validators=[MinValueValidator(-90.0), MaxValueValidator(90.0)]
+    )
+    latitude_max = models.FloatField(
+        default=13.9, validators=[MinValueValidator(-90.0), MaxValueValidator(90.0)]
+    )
+    longitude_min = models.FloatField(
+        default=2.7, validators=[MinValueValidator(-180.0), MaxValueValidator(180.0)]
+    )
+    longitude_max = models.FloatField(
+        default=14.7, validators=[MinValueValidator(-180.0), MaxValueValidator(180.0)]
+    )
+
     def __str__(self):
         return f"Options {self.id}: Project {self.project.name}"
