+++ offgridplanner/steps/views.py
@@ -1,8 +1,6 @@
 import os
 
-import pandas as pd
 from django.core.exceptions import PermissionDenied
-from django.forms import model_to_dict
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
 from django.shortcuts import redirect
@@ -17,9 +15,9 @@ from offgridplanner.optimization.supply.demand_estimation import ENTERPRISE_LIST
 from offgridplanner.optimization.supply.demand_estimation import LARGE_LOAD_KW_MAPPING
 from offgridplanner.optimization.supply.demand_estimation import LARGE_LOAD_LIST
 from offgridplanner.optimization.supply.demand_estimation import PUBLIC_SERVICE_LIST
+from offgridplanner.projects.forms import BoundForm  # Bachirou
 from offgridplanner.projects.forms import OptionForm
 from offgridplanner.projects.forms import ProjectForm
-from offgridplanner.projects.helpers import OUTPUT_KPIS
 from offgridplanner.projects.helpers import get_param_from_metadata
 from offgridplanner.projects.helpers import group_form_by_component
 from offgridplanner.projects.helpers import reorder_dict
@@ -49,6 +47,7 @@ STEP_LIST_RIBBON = [step for step in STEPS.values() if step != _("Calculating")]
 # @login_required()
 @require_http_methods(["GET", "POST"])
 def project_setup(request, proj_id=None):
+    global bound_form
     if proj_id is not None:
         project = get_object_or_404(Project, id=proj_id)
         if project.user != request.user:
@@ -62,14 +61,17 @@ def project_setup(request, proj_id=None):
         if project is not None:
             form = ProjectForm(instance=project)
             opts = OptionForm(instance=project.options)
+            bounds = BoundForm(instance=project.options)  # Added Bachirou
             context.update({"proj_id": project.id})
         else:
             form = ProjectForm(initial=get_param_from_metadata("default", "Project"))
             opts = OptionForm()
+            bounds = BoundForm()  # Bachirou
         context.update(
             {
                 "form": form,
                 "opts_form": opts,
+                "bounds_form": bounds,  # Bachirou added
                 # fields that should be rendered in left column (for use in template tags)
                 "left_col_fields": ["name", "n_days", "description"],
                 "max_days": max_days,
@@ -92,7 +94,10 @@ def project_setup(request, proj_id=None):
         else:
             form = ProjectForm(request.POST, instance=project)
             opts_form = OptionForm(request.POST, instance=project.options)
-        if form.is_valid() and opts_form.is_valid():
+            bound_form = BoundForm(
+                request.POST, instance=project.options
+            )  # Added Bachirou and 95 check
+        if form.is_valid() and opts_form.is_valid() and bound_form.is_valid():
             opts = opts_form.save()
             if project is None:
                 project = form.save(commit=False)
@@ -310,27 +315,12 @@ def calculating(request, proj_id=None):
 @require_http_methods(["GET"])
 def simulation_results(request, proj_id=None):
     step_id = list(STEPS.keys()).index("calculating") + 1
-
-    project = get_object_or_404(Project, id=proj_id)
-    opts = project.options
-    res = project.simulation.results
-    df = pd.Series(model_to_dict(res))
-
-    df = df.astype(float)
-    output_kpis = OUTPUT_KPIS.copy()
-
-    for kpi in output_kpis:
-        output_kpis[kpi]["value"] = df[kpi].round(1)
-
     return render(
         request,
         "pages/simulation_results.html",
         context={
             "proj_id": proj_id,
             "step_id": step_id,
-            "results": output_kpis,
-            "do_grid_optimization": opts.do_grid_optimization,
-            "do_supply_optimization": opts.do_es_design_optimization,
             "step_list": STEP_LIST_RIBBON,
         },
     )
